# [dotnet] [bidi] Modules as extensions