--- a/dotnet/src/webdriver/BiDi/BiDi.cs
+++ b/dotnet/src/webdriver/BiDi/BiDi.cs
@@ -18,6 +18,7 @@
 // </copyright>
 
 using System;
+using System.Collections.Concurrent;
 using System.Threading;
 using System.Threading.Tasks;
 using OpenQA.Selenium.BiDi.Communication;
@@ -28,18 +29,7 @@ public sealed class BiDi : IAsyncDisposable
 {
     private readonly Broker _broker;
 
-    private Session.SessionModule? _sessionModule;
-    private BrowsingContext.BrowsingContextModule? _browsingContextModule;
-    private Browser.BrowserModule? _browserModule;
-    private Network.NetworkModule? _networkModule;
-    private Input.InputModule? _inputModule;
-    private Script.ScriptModule? _scriptModule;
-    private Log.LogModule? _logModule;
-    private Storage.StorageModule? _storageModule;
-    private WebExtension.WebExtensionModule? _webExtensionModule;
-    private Emulation.EmulationModule? _emulationModule;
-
-    private readonly object _moduleLock = new();
+    private readonly ConcurrentDictionary<Type, Module> _modules = [];
 
     private BiDi(string url)
     {
@@ -48,134 +38,29 @@ public sealed class BiDi : IAsyncDisposable
         _broker = new Broker(this, uri);
     }
 
-    internal Session.SessionModule SessionModule
-    {
-        get
-        {
-            if (_sessionModule is not null) return _sessionModule;
-            lock (_moduleLock)
-            {
-                _sessionModule ??= Module.Create<Session.SessionModule>(this, _broker);
-            }
-            return _sessionModule;
-        }
-    }
+    internal Session.SessionModule SessionModule => AsModule<Session.SessionModule>();
 
-    public BrowsingContext.BrowsingContextModule BrowsingContext
-    {
-        get
-        {
-            if (_browsingContextModule is not null) return _browsingContextModule;
-            lock (_moduleLock)
-            {
-                _browsingContextModule ??= Module.Create<BrowsingContext.BrowsingContextModule>(this, _broker);
-            }
-            return _browsingContextModule;
-        }
-    }
+    public BrowsingContext.BrowsingContextModule BrowsingContext => AsModule<BrowsingContext.BrowsingContextModule>();
 
-    public Browser.BrowserModule Browser
-    {
-        get
-        {
-            if (_browserModule is not null) return _browserModule;
-            lock (_moduleLock)
-            {
-                _browserModule ??= Module.Create<Browser.BrowserModule>(this, _broker);
-            }
-            return _browserModule;
-        }
-    }
+    public Browser.BrowserModule Browser => AsModule<Browser.BrowserModule>();
 
-    public Network.NetworkModule Network
-    {
-        get
-        {
-            if (_networkModule is not null) return _networkModule;
-            lock (_moduleLock)
-            {
-                _networkModule ??= Module.Create<Network.NetworkModule>(this, _broker);
-            }
-            return _networkModule;
-        }
-    }
+    public Network.NetworkModule Network => AsModule<Network.NetworkModule>();
 
-    internal Input.InputModule InputModule
-    {
-        get
-        {
-            if (_inputModule is not null) return _inputModule;
-            lock (_moduleLock)
-            {
-                _inputModule ??= Module.Create<Input.InputModule>(this, _broker);
-            }
-            return _inputModule;
-        }
-    }
+    internal Input.InputModule InputModule => AsModule<Input.InputModule>();
 
-    public Script.ScriptModule Script
-    {
-        get
-        {
-            if (_scriptModule is not null) return _scriptModule;
-            lock (_moduleLock)
-            {
-                _scriptModule ??= Module.Create<Script.ScriptModule>(this, _broker);
-            }
-            return _scriptModule;
-        }
-    }
+    public Script.ScriptModule Script => AsModule<Script.ScriptModule>();
 
-    public Log.LogModule Log
-    {
-        get
-        {
-            if (_logModule is not null) return _logModule;
-            lock (_moduleLock)
-            {
-                _logModule ??= Module.Create<Log.LogModule>(this, _broker);
-            }
-            return _logModule;
-        }
-    }
+    public Log.LogModule Log => AsModule<Log.LogModule>();
 
-    public Storage.StorageModule Storage
-    {
-        get
-        {
-            if (_storageModule is not null) return _storageModule;
-            lock (_moduleLock)
-            {
-                _storageModule ??= Module.Create<Storage.StorageModule>(this, _broker);
-            }
-            return _storageModule;
-        }
-    }
+    public Storage.StorageModule Storage => AsModule<Storage.StorageModule>();
 
-    public WebExtension.WebExtensionModule WebExtension
-    {
-        get
-        {
-            if (_webExtensionModule is not null) return _webExtensionModule;
-            lock (_moduleLock)
-            {
-                _webExtensionModule ??= Module.Create<WebExtension.WebExtensionModule>(this, _broker);
-            }
-            return _webExtensionModule;
-        }
-    }
+    public WebExtension.WebExtensionModule WebExtension => AsModule<WebExtension.WebExtensionModule>();
 
-    public Emulation.EmulationModule Emulation
+    public Emulation.EmulationModule Emulation => AsModule<Emulation.EmulationModule>();
+
+    public TModule AsModule<TModule>() where TModule : Module, new()
     {
-        get
-        {
-            if (_emulationModule is not null) return _emulationModule;
-            lock (_moduleLock)
-            {
-                _emulationModule ??= Module.Create<Emulation.EmulationModule>(this, _broker);
-            }
-            return _emulationModule;
-        }
+        return (TModule)_modules.GetOrAdd(typeof(TModule), _ => Module.Create<TModule>(this, _broker));
     }
 
     public Task<Session.StatusResult> StatusAsync()