--- a/dotnet/src/webdriver/BiDi/Browser/BrowserModule.cs
+++ b/dotnet/src/webdriver/BiDi/Browser/BrowserModule.cs
@@ -18,9 +18,6 @@
 // </copyright>
 
 using OpenQA.Selenium.BiDi.Communication;
-using System.Collections.Generic;
-using System.Text.Json;
-using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
 namespace OpenQA.Selenium.BiDi.Browser;
@@ -76,25 +73,4 @@ public sealed class BrowserModule : Module
 
         return await Broker.ExecuteCommandAsync<SetDownloadBehaviorCommand, EmptyResult>(new SetDownloadBehaviorCommand(@params), options, JsonContext).ConfigureAwait(false);
     }
-
-    protected internal override JsonSerializerContext ConfigureJson(JsonSerializerOptions options)
-    {
-        return new BrowserModuleJsonSerializerContext(options);
-    }
 }
-
-[JsonSerializable(typeof(Command))]
-[JsonSerializable(typeof(EmptyResult))]
-
-[JsonSerializable(typeof(CloseCommand))]
-[JsonSerializable(typeof(CreateUserContextCommand))]
-[JsonSerializable(typeof(GetUserContextsCommand))]
-[JsonSerializable(typeof(GetUserContextsResult))]
-[JsonSerializable(typeof(RemoveUserContextCommand))]
-[JsonSerializable(typeof(GetClientWindowsCommand))]
-[JsonSerializable(typeof(GetClientWindowsResult))]
-[JsonSerializable(typeof(SetDownloadBehaviorCommand))]
-[JsonSerializable(typeof(UserContextInfo))]
-[JsonSerializable(typeof(IReadOnlyList<UserContextInfo>))]
-[JsonSerializable(typeof(IReadOnlyList<ClientWindowInfo>))]
-internal partial class BrowserModuleJsonSerializerContext : JsonSerializerContext;