### **User description**
Contributes to https://github.com/SeleniumHQ/selenium/issues/15329

### 🔗 Related Issues
<!-- Example: Fixes #1234 or Closes #5678 -->
<!-- If the reason for this PR is not obvious, consider creating an issue for it first -->

### 💥 What does this PR do?
<!-- Describe what this change includes and how it works -->

### 🔧 Implementation Notes
<!--- Why did you implement it this way? -->
<!--- What alternatives to this approach did you consider? -->

### 💡 Additional Considerations
<!--- Are there any decisions that need to be made before accepting this PR? -->
<!--- Is there any follow-on work that needs to be done? (e.g., docs, tests, etc.) -->

### 🔄 Types of changes
<!-- ✂️ Please delete anything that doesn't apply -->
- Cleanup (formatting, renaming)
- Bug fix (backwards compatible)
- New feature (non-breaking change which adds functionality *and tests!*)
- Breaking change (fix or feature that would cause existing functionality to change)


___

### **PR Type**
Enhancement


___

### **Description**
- Refactor BiDi modules to use extension pattern

- Centralize JSON serialization configuration in BiDi class

- Replace individual module properties with generic AsModule method

- Update all module constructors to parameterless design


___

### Diagram Walkthrough


```mermaid
flowchart LR
  A["BiDi Class"] --> B["AsModule<T>() Method"]
  B --> C["ConcurrentDictionary<Type, Module>"]
  C --> D["Module.Create<T>()"]
  D --> E["Individual Modules"]
  A --> F["Centralized JsonSerializerOptions"]
  F --> G["BiDiJsonSerializerContext"]
  G --> E
```



<details> <summary><h3> File Walkthrough</h3></summary>

<table><thead><tr><th></th><th align="left">Relevant files</th></tr></thead><tbody><tr><td><strong>Enhancement</strong></td><td><details><summary>13 files</summary><table>
<tr>
  <td><strong>BiDi.cs</strong><dd><code>Centralize JSON context and implement module extensions</code>&nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-3f77d71f94e2a47ea90b74f6d37b5a5b94dc83b2f558b30ab070588b9028c2f6">+46/-129</a></td>

</tr>

<tr>
  <td><strong>Module.cs</strong><dd><code>Add factory method and JSON context properties</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-b45f63b7ea084aed81f19bba490ac80626f908cecea9448d35db3d4fbc663ddc">+22/-2</a>&nbsp; &nbsp; </td>

</tr>

<tr>
  <td><strong>Broker.cs</strong><dd><code>Update command execution with JSON context parameter</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-5f7fc0e0b816bad29734139ccd27ebb0ff8809c8f49de7827eab7c6c7e8f9669">+22/-58</a>&nbsp; </td>

</tr>

<tr>
  <td><strong>BrowserModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-8e4bcc00ef63cf737c6e26397ac40f1e18d960e4ff47589fee1bae1b9fe8cdc7">+10/-10</a>&nbsp; </td>

</tr>

<tr>
  <td><strong>BrowsingContextModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-1f04192c50706a21f9dfdc1b79916855ccf2aba6c2c895811f158eeac915680c">+41/-41</a>&nbsp; </td>

</tr>

<tr>
  <td><strong>SessionModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-31b1c3c5b298bcc68ac630c0af2fe9ecd548a737be8b5fa14574317f6eff3ff3">+6/-6</a>&nbsp; &nbsp; &nbsp; </td>

</tr>

<tr>
  <td><strong>NetworkModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-ddd0c9f9565ffffc6a3ce489431b76319cc38c727006605f7dc6345b2c0f7c81">+25/-25</a>&nbsp; </td>

</tr>

<tr>
  <td><strong>ScriptModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-da8cf630954e0732aa2b7f257aca8b49117eeceae343e373e2ba5914e639b6d2">+12/-12</a>&nbsp; </td>

</tr>

<tr>
  <td><strong>LogModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-fdb6bcf97d10b7240dada7ced38218b43fcb0b8ba716b273dd9c60e17c286e3f">+3/-3</a>&nbsp; &nbsp; &nbsp; </td>

</tr>

<tr>
  <td><strong>StorageModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-c2832a3991518e29f9a8f706cbe7e7c1d5ec4a007fbb57da280bb1be9fc7b2c2">+4/-4</a>&nbsp; &nbsp; &nbsp; </td>

</tr>

<tr>
  <td><strong>InputModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-f98f384929541b30f9b87fb08f96ead9181d6b14a2353f5485ba985ed7cbd672">+4/-4</a>&nbsp; &nbsp; &nbsp; </td>

</tr>

<tr>
  <td><strong>EmulationModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-326495cc635fdb9686d337228f031a251b2aaed55960d7096b1345e0f1c267ff">+10/-11</a>&nbsp; </td>

</tr>

<tr>
  <td><strong>WebExtensionModule.cs</strong><dd><code>Remove constructor parameter and use JsonContext</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/16392/files#diff-0eaa7ea3946fcd852a8db4e125d1bd98c674f8ce8489cc0279f452dca8f30a25">+3/-3</a>&nbsp; &nbsp; &nbsp; </td>

</tr>
</table></details></td></tr></tr></tbody></table>

</details>

___