@nvborisenko 2025-10-06
I don't want to believe that injecting custom `TypeInfoResolver` is not possible.