@nvborisenko 2025-10-07
Now each module owns `JsonSerializerContext`, and I see 2 issues:

### **Performance degradation (memory)**

Before:
```
| Method                         | Mean     | Error   | StdDev  | Gen0     | Gen1     | Gen2     | Allocated |
|------------------------------- |---------:|--------:|--------:|---------:|---------:|---------:|----------:|
| SeleniumCaptureScreenshotAsync | 128.1 ms | 6.83 ms | 4.52 ms | 250.0000 | 250.0000 | 250.0000 |   1.65 MB |
```

After:
```
| Method                         | Mean     | Error   | StdDev  | Gen0     | Gen1     | Gen2     | Allocated |
|------------------------------- |---------:|--------:|--------:|---------:|---------:|---------:|----------:|
| SeleniumCaptureScreenshotAsync | 128.5 ms | 9.98 ms | 5.94 ms | 500.0000 | 500.0000 | 500.0000 |   2.31 MB |
```

### **Cross-module type referencing**

If we define json context per module, then we also have to define all types this module is referencing to. It is nightmare! We should track references somehow, and it is easy to make a mistake.

### How to move on?
I think we can use one big shared json context by all modules. External modules will define their own context.