@nvborisenko 2025-10-07
@RenderMichael I am confident with proposed changes, please review.

- `BiDi` is owner of `JsonOptions/Context`
- Built-in modules reuse `JsonContext`
- External module may create own `JsonContext` and propagate to commands